ml/flaskapi/requirements.txt
absl-py==2.0.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
dm-tree==0.1.8
filelock==3.12.4
Flask==3.0.0
fsspec==2023.9.2
google-api-core==2.12.0
google-auth==2.23.2
google-cloud==0.34.0
google-cloud-core==2.3.3
google-cloud-storage==2.11.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
h5py==3.9.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
keras-core==0.1.7
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mpmath==1.3.0
namex==0.0.7
networkx==3.1
numpy==1.26.0
Pillow==10.0.1
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.16.1
requests==2.31.0
rich==13.6.0
rsa==4.9
sympy==1.12
torch==2.1.0
typing_extensions==4.8.0
urllib3==2.0.6
Werkzeug==3.0.0
